requirements.txt
vosk
pyaudio
spotipy
openai
python-dotenv
colorlog
pytest
